Trait s2n_quic_core::ack::set::Set

pub trait Set {
    fn contains(&self, packet_number: PacketNumber) -> bool;
    fn smallest(&self) -> PacketNumber;
    fn largest(&self) -> PacketNumber;

    fn as_range(&self) -> PacketNumberRange  { ... }
}

A set of Acknowledgments

The implementation of the set is allowed to store packet numbers in an arbitrary form.

Required Methods

fn contains(&self, packet_number: PacketNumber) -> bool

Returns whether the Set contains a given packet number

fn smallest(&self) -> PacketNumber

Smallest packet number in the set

fn largest(&self) -> PacketNumber

Largest packet number in the set

Provided Methods

fn as_range(&self) -> PacketNumberRange

Returns the range of packet numbers contained in the set