Trait s2n_quic_core::connection::id::Generator

pub trait Generator {
    fn generate(&mut self, connection_info: &ConnectionInfo<'_>) -> LocalId;

    fn lifetime(&self) -> Option<Duration> { ... }
}

A generator for a connection ID format

Required Methods

fn generate(&mut self, connection_info: &ConnectionInfo<'_>) -> LocalId

Generates a connection ID.

Connection IDs MUST NOT contain any information that can be used by an external observer (that is, one that does not cooperate with the issuer) to correlate them with other connection IDs for the same connection.

Each call to generate should produce a unique Connection ID, otherwise the endpoint may terminate.

Provided Methods

fn lifetime(&self) -> Option<Duration>

The maximum amount of time each generated connection ID should be used for. By default there is no maximum, though connection IDs may be retired due to rotation requirements or peer requests.