Trait s2n_quic_core::connection::id::Validator

pub trait Validator {
    fn validate(
        &self,
        connection_info: &ConnectionInfo<'_>,
        buffer: &[u8]
    ) -> Option<usize>;
}

A validator for a connection ID format

Required Methods

fn validate(
&self,
connection_info: &ConnectionInfo<'_>,
buffer: &[u8]
) -> Option<usize>

Validates a connection ID from a buffer

Implementations should handle situations where the buffer will include extra data after the connection ID.

Returns the length of the connection id if successful, otherwise None is returned.