Struct s2n_quic_core::crypto::application::KeySet

pub struct KeySet<K> { /* private fields */ }

Implementations

impl<K: OneRttKey> KeySet<K>

pub fn new(crypto: K, limits: Limits) -> Self

pub fn set_derivation_timer(&mut self, timestamp: Timestamp)

Set the timer to derive a new key after timestamp

pub fn key_update_in_progress(&self) -> bool

Returns whether there is a key update in progress.

pub fn decrypt_packet<'a>(
&mut self,
packet: EncryptedShort<'a>,
largest_acknowledged_packet_number: PacketNumber,
pto: Timestamp
) -> Result<(CleartextShort<'a>, Option<u16>), ProcessingError>

Passes the key for the the requested phase to a callback function. Integrity limits are enforced.

Returns the decrypted packet and generation if the key phase was rotated.

pub fn encryption_phase(&self) -> KeyPhase

This is the KeyPhase that should be used to encrypt a given packet.

pub fn encrypt_packet<'a, F>(
&mut self,
buffer: EncoderBuffer<'a>,
f: F
) -> Result<(ProtectedPayload<'a>, EncoderBuffer<'a>), PacketEncodingError<'a>>where
F: FnOnce(EncoderBuffer<'a>, &K, KeyPhase) -> Result<(ProtectedPayload<'a>, EncoderBuffer<'a>), PacketEncodingError<'a>>,

pub fn on_timeout(&mut self, timestamp: Timestamp)

pub fn key_phase(&self) -> KeyPhase

pub fn active_key(&self) -> &Key<K>

pub fn active_key_mut(&mut self) -> &mut Key<K>

pub fn cipher_suite(&self) -> CipherSuite