Module s2n_quic_core::datagram::traits

Structs

ConnectionInfo

ConnectionInfo contains the peer’s limit on the size of datagrams they accept

PreConnectionInfo

PreConnectionInfo will contain information needed to determine whether or not a provider will accept datagrams.

Enums

WriteError

Traits

Endpoint

The datagram endpoint trait provides a way to implement custom unreliable datagram sending and receiving logic. The Sender type should be implemented for custom sending behavior, and the Receiver type should be implemented for custom receiving behavior.

Packet

A packet will be available during the on_transmit callback. Use the methods defined here to interrogate the packet struct and write datagrams to the packet.

Receiver

Allows users to configure the behavior of receiving datagrams.

Sender

Allows users to configure the behavior of sending datagrams.