Trait s2n_quic_core::datagram::traits::Endpoint

pub trait Endpoint: 'static + Send {
    type Sender: Sender;
    type Receiver: Receiver;

    fn create_connection(
        &mut self,
        info: &ConnectionInfo
    ) -> (Self::Sender, Self::Receiver);
    fn max_datagram_frame_size(&self, info: &PreConnectionInfo) -> u64;
}

The datagram endpoint trait provides a way to implement custom unreliable datagram sending and receiving logic. The Sender type should be implemented for custom sending behavior, and the Receiver type should be implemented for custom receiving behavior.

Required Associated Types

type Sender: Sender

type Receiver: Receiver

Required Methods

fn create_connection(
&mut self,
info: &ConnectionInfo
) -> (Self::Sender, Self::Receiver)

fn max_datagram_frame_size(&self, info: &PreConnectionInfo) -> u64

Returns the maximum datagram frame size the provider is willing to accept