Module s2n_quic_core::event::api

This module contains events that are emitted to the Subscriber

Structs

AckProcessedDeprecated

Events related to ACK processing

AckRangeReceived

ACK range was received

ActivePathUpdated

Active path was updated

ApplicationProtocolInformation

Application level protocol

BbrStateChanged

The BBR state has changed

Congestion

Congestion (ECN or packet loss) has occurred

ConnectionClosed

Connection closed

ConnectionId

ConnectionIdUpdated

ConnectionId updated

ConnectionInfo

ConnectionMeta

ConnectionMigrationDenied

ConnectionStarted

Connection started

DatagramDropped

Datagram dropped by a connection

DatagramReceived

Datagram received by a connection

DatagramSent

Datagram sent by a connection

DeliveryRateSampled

A new delivery rate sample has been generated Note: This event is only recorded for congestion controllers that support bandwidth estimates, such as BBR

DuplicatePacket

Duplicate packet received

EcnCounts

EcnStateChanged

EndpointConnectionAttemptFailed

EndpointDatagramDropped

Datagram dropped by the endpoint

EndpointDatagramReceived

Datagram received by the endpoint

EndpointDatagramSent

Datagram sent by the endpoint

EndpointMeta

EndpointPacketReceived

Packet was received by the endpoint

EndpointPacketSent

Packet was sent by the endpoint

FrameReceived

Frame was received

FrameSent

Frame was sent

HandshakeStatusUpdated

KeepAliveTimerExpired

KeySpaceDiscarded

KeyUpdate

Crypto key updated

MtuUpdated

The maximum transmission unit (MTU) for the path has changed

PacingRateUpdated

The pacing rate has been updated

PacketDropped

Packet was dropped with the given reason

PacketLost

Packet was lost

PacketReceived

Packet was received by a connection

PacketSent

Packet was sent by a connection

Path

PathChallengeUpdated

Path challenge updated

PathCreated

A new path was created

PlatformEventLoopSleep

PlatformEventLoopWakeup

PlatformFeatureConfigured

Emitted when a platform feature is configured

PlatformRx

Emitted when the platform receives at least one packet

PlatformRxError

Emitted when the platform returns an error while receiving datagrams

PlatformTx

Emitted when the platform sends at least one packet

PlatformTxError

Emitted when the platform returns an error while sending datagrams

PreferredAddress

RateSample

A bandwidth delivery rate estimate with associated metadata

RecoveryMetrics

Recovery metrics updated

RxAckRangeDropped

Ack range for received packets was dropped due to space constraints

RxStreamProgress

ServerNameInformation

Server Name was negotiated for the connection

SlowStartExited

The slow start congestion controller state has been exited

TlsClientHello

TlsServerHello

TransportParameters

TransportParametersReceived

Transport parameters received by connection

TxStreamProgress

VersionInformation

QUIC version

Enums

AckActionDeprecated

BbrState

CipherSuite

CongestionSource

The source that caused a congestion event

DatagramDropReason

DuplicatePacketError

EcnState

The current state of the ECN controller for the path

EndpointType

An endpoint may be either a Server or a Client

Frame

HandshakeStatus

Events tracking the progress of handshake status

KeySpace

KeyType

MigrationDenyReason

MtuUpdatedCause

The reason the MTU was updated

PacketDropReason

PacketHeader

PathChallengeStatus

PlatformFeatureConfiguration

RetryDiscardReason

SlowStartExitCause

The reason the slow start congestion controller state has been exited

SocketAddress

StreamType

Subject

A context from which the event is being emitted

Traits

Subscriber

Allows for events to be subscribed to