Struct s2n_quic_core::frame::crypto::Crypto

pub struct Crypto<Data> {
    pub offset: VarInt,
    pub data: Data,
}

Fields
offset: VarInt

A variable-length integer specifying the byte offset in the stream for the data in this CRYPTO frame.

data: Data

The cryptographic message data.

Implementations

impl<Data> Crypto<Data>

pub const fn tag(&self) -> u8

pub fn map_data<F: FnOnce(Data) -> Out, Out>(self, map: F) -> Crypto<Out>

Converts the crypto data from one type to another

impl<Data: EncoderValue> Crypto<Data>

pub fn try_fit(&self, capacity: usize) -> Result<usize, FitError>

Tries to fit the frame into the provided capacity

If ok, the new payload length is returned, otherwise the frame cannot fit.