Struct s2n_quic_core::havoc::EncoderBuffer

pub struct EncoderBuffer<'a> { /* private fields */ }

EncoderBuffer is a buffer for writing to a mutable slice

Implementations

impl<'a> EncoderBuffer<'a>

pub fn new(bytes: &'a mut [u8]) -> EncoderBuffer<'a>

Creates a new EncoderBuffer

pub fn set_position(&mut self, position: usize)

Sets the write cursor to a new position

Panics

Panics when position > capacity

pub fn advance_position(&mut self, offset: usize)

Advances the write cursor by offset

Panics

Panics when position > capacity

pub fn split_off(self) -> (&'a mut [u8], &'a mut [u8])

Splits off the used buffer from the remaining bytes

pub fn split_mut(&mut self) -> (&mut [u8], &mut [u8])

Splits the used buffer from the remaining bytes

pub fn as_mut_slice(&mut self) -> &mut [u8]

Returns the written bytes as a mutable slice