Trait s2n_quic_core::havoc::EncoderValue

pub trait EncoderValue: Sized {
    fn encode<E>(&self, encoder: &mut E)
    where
        E: Encoder;

    fn encode_mut<E>(&mut self, encoder: &mut E)
    where
        E: Encoder,
    { ... }
    fn encoding_size(&self) -> usize { ... }
    fn encoding_size_for_encoder<E>(&self, encoder: &E) -> usize
    where
        E: Encoder,
    { ... }
    fn encode_with_len_prefix<Len, E>(&self, encoder: &mut E)
    where
        Len: TryFrom<usize> + EncoderValue,
        E: Encoder,
        Self: Sized,
        <Len as TryFrom<usize>>::Error: Debug,
    { ... }
}

Required Methods

fn encode<E>(&self, encoder: &mut E)where
E: Encoder,

Encodes the value into the encoder

Provided Methods

fn encode_mut<E>(&mut self, encoder: &mut E)where
E: Encoder,

Encodes the value into the encoder, while potentially mutating the value itself

fn encoding_size(&self) -> usize

Returns the encoding size with no buffer constrains

fn encoding_size_for_encoder<E>(&self, encoder: &E) -> usizewhere
E: Encoder,

Returns the encoding size for the given encoder’s capacity

fn encode_with_len_prefix<Len, E>(&self, encoder: &mut E)where
Len: TryFrom<usize> + EncoderValue,
E: Encoder,
Self: Sized,
<Len as TryFrom<usize>>::Error: Debug,

Encodes the value into the encoder with a prefix of Len