Trait s2n_quic_core::io::tx::Message

pub trait Message {
    type Handle: Handle;

    fn path_handle(&self) -> &Self::Handle;
    fn ecn(&mut self) -> ExplicitCongestionNotification;
    fn delay(&mut self) -> Duration;
    fn ipv6_flow_label(&mut self) -> u32;
    fn can_gso(&self, segment_len: usize, segment_count: usize) -> bool;
    fn write_payload(
        &mut self,
        buffer: PayloadBuffer<'_>,
        gso_offset: usize
    ) -> Result<usize, Error>;
}

Abstraction over a message to be sent on a socket

Instead of a concrete struct with eagerly evaluated fields, using trait callbacks ensure messages only need to compute what the actual transmission queue requires. For example, if the transmission queue cannot set ECN markings, it will not call the Message::ecn function.

Required Associated Types

type Handle: Handle

Required Methods

fn path_handle(&self) -> &Self::Handle

Returns the path handle on which this message should be sent

fn ecn(&mut self) -> ExplicitCongestionNotification

Returns the ECN markings for the message

fn delay(&mut self) -> Duration

Returns the Duration for which the message will be delayed.

This is used in scenarios where packets need to be paced.

fn ipv6_flow_label(&mut self) -> u32

Returns the IPv6 flow label for the message

fn can_gso(&self, segment_len: usize, segment_count: usize) -> bool

Returns true if the packet can be used in a GSO packet

fn write_payload(
&mut self,
buffer: PayloadBuffer<'_>,
gso_offset: usize
) -> Result<usize, Error>

Writes the payload of the message to an output buffer