Trait s2n_quic_core::io::tx::Queue

pub trait Queue {
    type Entry: Entry<Handle = Self::Handle>;
    type Handle: Handle;

    const SUPPORTS_ECN: bool = false;
    const SUPPORTS_PACING: bool = false;
    const SUPPORTS_FLOW_LABELS: bool = false;

    fn push<M: Message<Handle = Self::Handle>>(
        &mut self,
        message: M
    ) -> Result<Outcome, Error>;
    fn as_slice_mut(&mut self) -> &mut [Self::Entry];
    fn capacity(&self) -> usize;
    fn len(&self) -> usize;

    fn has_capacity(&self) -> bool { ... }
    fn is_empty(&self) -> bool { ... }
}

A structure capable of queueing and transmitting messages

Required Associated Types

type Entry: Entry<Handle = Self::Handle>

type Handle: Handle

Provided Associated Constants

const SUPPORTS_ECN: bool = false

Set to true if the queue supports setting ECN markings

const SUPPORTS_PACING: bool = false

Set to true if the queue supports pacing of sending messages

const SUPPORTS_FLOW_LABELS: bool = false

Set to true if the queue supports setting IPv6 flow labels

Required Methods

fn push<M: Message<Handle = Self::Handle>>(
&mut self,
message: M
) -> Result<Outcome, Error>

Pushes a message into the transmission queue

The index of the message is returned to enable further operations to be performed, e.g. encryption.

fn as_slice_mut(&mut self) -> &mut [Self::Entry]

Returns the pending messages as a mutable slice

fn capacity(&self) -> usize

Returns the number of remaining datagrams that can be transmitted

fn len(&self) -> usize

Returns the number of pending datagrams to be transmitted

Provided Methods

fn has_capacity(&self) -> bool

Returns true if the queue will accept additional transmissions

fn is_empty(&self) -> bool

Returns true if there are no pending datagrams to be transmitted