Module s2n_quic_core::packet::number

Re-exports

pub use map::Map;

Modules

map

Structs

PacketNumber

Contains a fully-decoded packet number in a given space

PacketNumberLen

A fully-decoded and unprotected packet number length

PacketNumberRange

An inclusive range of PacketNumbers

ProtectedPacketNumber

A placeholder for PacketNumber values.

SlidingWindow

TruncatedPacketNumber

A truncated packet number, which is derived from the largest acknowledged packet number

Enums

PacketNumberSpace

Contains all of the available packet spaces for QUIC packets

SlidingWindowError