Trait s2n_quic_core::path::Handle

pub trait Handle: 'static + Copy + Send + Debug {
    fn from_remote_address(remote_addr: RemoteAddress) -> Self;
    fn remote_address(&self) -> RemoteAddress;
    fn local_address(&self) -> LocalAddress;
    fn eq(&self, other: &Self) -> bool;
    fn strict_eq(&self, other: &Self) -> bool;
}

An interface for an object that represents a unique path between two endpoints

Required Methods

fn from_remote_address(remote_addr: RemoteAddress) -> Self

Creates a Handle from a RemoteAddress

fn remote_address(&self) -> RemoteAddress

Returns the remote address for the given handle

fn local_address(&self) -> LocalAddress

Returns the local address for the given handle

fn eq(&self, other: &Self) -> bool

Returns true if the two handles are equal from a network perspective

This function is used to determine if a connection has migrated to another path.

fn strict_eq(&self, other: &Self) -> bool

Returns true if the two handles are strictly equal to each other, i.e. byte-for-byte.