Struct s2n_quic_core::stream::ops::Request

pub struct Request<'a> {
    pub tx: Option<Request<'a>>,
    pub rx: Option<Request<'a>>,
}

A request made on a stream

Fields
tx: Option<Request<'a>>

The tx options of the request

rx: Option<Request<'a>>

The rx options of the request

Implementations

impl<'a> Request<'a>

pub fn send(&mut self, chunks: &'a mut [Bytes]) -> &mut Self

Requests a slice of chunks to be sent on the tx stream

pub fn reset(&mut self, error: Error) -> &mut Self

Resets the tx stream with an error code

pub fn flush(&mut self) -> &mut Self

Flushes any pending tx data to be ACKed before unblocking

pub fn finish(&mut self) -> &mut Self

Marks the tx stream as finished (e.g. no more data will be sent)

pub fn receive(&mut self, chunks: &'a mut [Bytes]) -> &mut Self

Requests data on the rx stream to be received into the provided slice of chunks

pub fn stop_sending(&mut self, error: Error) -> &mut Self

Requests the peer to stop sending data on the rx stream

pub fn with_watermark(&mut self, low: usize, high: usize) -> &mut Self

Sets the watermarks for the rx stream

pub fn with_low_watermark(&mut self, low: usize) -> &mut Self

Sets the low watermark for the rx stream

If the watermark is set to 0, the caller will be notified as soon as there is data available on the stream.

If the watermark is greater than 0, the caller will be notified as soon as there is at least low bytes available to be read. Note that the stream may be woken earlier.

pub fn with_high_watermark(&mut self, high: usize) -> &mut Self

Sets the high watermark for the rx stream

The stream ensures that all the received data will not exceed the watermark amount. This can be useful for receiving at most n bytes.

pub fn detach_tx(&mut self) -> &mut Self

pub fn detach_rx(&mut self) -> &mut Self