Struct s2n_quic_core::time::timer::Timer

pub struct Timer { /* private fields */ }

A timer that does not trigger an update in a timer list. These are usually owned by individual components and needs to be explicitly polled.

Note: The timer doesn’t implement Copy to ensure it isn’t accidentally moved and have the expiration discarded.

Implementations

impl Timer

pub fn set(&mut self, time: Timestamp)

Sets the timer to expire at the given timestamp

pub fn cancel(&mut self)

Cancels the timer. After cancellation, a timer will no longer report as expired.

pub fn is_expired(&self, current_time: Timestamp) -> bool

Returns true if the timer has expired

pub fn is_armed(&self) -> bool

Returns true if the timer is armed

pub fn poll_expiration(&mut self, current_time: Timestamp) -> Poll<()>

Notifies the timer of the current time. If the timer’s expiration occurs before the current time, it will be cancelled. The method returns whether the timer was expired and had been cancelled.