Trait s2n_quic_core::time::timer::Provider

pub trait Provider {
    fn timers<Q: Query>(&self, query: &mut Q) -> Result;

    fn next_expiration(&self) -> Option<Timestamp> { ... }
    fn is_armed(&self) -> bool { ... }
    fn armed_timer_count(&self) -> usize { ... }
    fn for_each_timer<F: FnMut(&Timer) -> Result>(&self, f: F) { ... }
}

A trait for a components that owns at least one timer

Required Methods

fn timers<Q: Query>(&self, query: &mut Q) -> Result

Notifies the query of any timers owned by the provider

The provider should also delegate to subcomponents that own timers as well.

Provided Methods

fn next_expiration(&self) -> Option<Timestamp>

Returns the next Timestamp at which the earliest timer is armed in the provider

fn is_armed(&self) -> bool

Returns true if there are any timers armed

fn armed_timer_count(&self) -> usize

Counts the number of armed timers in the provider

fn for_each_timer<F: FnMut(&Timer) -> Result>(&self, f: F)

Iterates over each timer in the provider and calls the provided function