Struct s2n_quic_core::transport::error::Error

pub struct Error {
    pub code: Code,
    pub frame_type: VarInt,
    pub reason: &'static str,
}

Transport Errors are 62-bit unsigned integer values indicating a QUIC transport error has occurred, as defined in QUIC Transport RFC.

Fields
code: Code

A 62-bit unsigned integer value indicating the error that occurred

frame_type: VarInt

If this error was caused by a particular QUIC frame, frame_type will contain the Frame Type as defined in QUIC Transport RFC.

reason: &'static str

Additional information about the error that occurred

Implementations

impl Error

pub const fn new(code: VarInt) -> Self

Creates a new Error

pub const fn with_frame_type(self, frame_type: VarInt) -> Self

Updates the Error with the specified frame_type

pub const fn with_reason(self, reason: &'static str) -> Self

Updates the Error with the specified reason

impl Error

pub const NO_ERROR: Self = _

An endpoint uses this with CONNECTION_CLOSE to signal that the connection is being closed abruptly in the absence of any error

pub const INTERNAL_ERROR: Self = _

The endpoint encountered an internal error and cannot continue with the connection.

pub const CONNECTION_REFUSED: Self = _

The server refused to accept a new connection.

pub const FLOW_CONTROL_ERROR: Self = _

An endpoint received more data than it permitted in its advertised data limits.

pub const STREAM_LIMIT_ERROR: Self = _

An endpoint received a frame for a stream identifier that exceeded its advertised stream limit for the corresponding stream type.

pub const STREAM_STATE_ERROR: Self = _

An endpoint received a frame for a stream that was not in a state that permitted that frame.

pub const FINAL_SIZE_ERROR: Self = _

An endpoint received a STREAM frame containing data that exceeded the previously established final size.

pub const FRAME_ENCODING_ERROR: Self = _

An endpoint received a frame that was badly formatted.

pub const TRANSPORT_PARAMETER_ERROR: Self = _

An endpoint received transport parameters that were badly formatted.

pub const CONNECTION_ID_LIMIT_ERROR: Self = _

The number of connection IDs provided by the peer exceeds the advertised active_connection_id_limit.

pub const PROTOCOL_VIOLATION: Self = _

An endpoint detected an error with protocol compliance that was not covered by more specific error codes.

pub const INVALID_TOKEN: Self = _

A server received a client Initial that contained an invalid Token field.

pub const APPLICATION_ERROR: Self = _

The application or application protocol caused the connection to be closed.

pub const CRYPTO_BUFFER_EXCEEDED: Self = _

An endpoint has received more data in CRYPTO frames than it can buffer.

pub const KEY_UPDATE_ERROR: Self = _

An endpoint detected errors in performing key updates.

pub const AEAD_LIMIT_REACHED: Self = _

An endpoint has reached the confidentiality or integrity limit for the AEAD algorithm used by the given connection.

pub fn description(&self) -> Option<&'static str>

impl Error

pub const fn crypto_error(code: u8) -> Self

Creates a crypto-level TransportError from a TLS alert code.

pub fn try_into_crypto_error(self) -> Option<CryptoError>

If the Error contains a CryptoError, it is returned

impl Error

pub const fn application_error(code: VarInt) -> Self

Creates an application-level Error