Module s2n_quic_core::transport::parameters

Structs

AckDelayExponent

ActiveConnectionIdLimit

DatagramLimits

DisabledParameter

Struct for marking a field as disabled for a given endpoint type

InitialFlowControlLimits

InitialMaxData

InitialMaxStreamDataBidiLocal

InitialMaxStreamDataBidiRemote

InitialMaxStreamDataUni

InitialMaxStreamsBidi

InitialMaxStreamsUni

InitialSourceConnectionId

InitialStreamLimits

Associated flow control limits from a set of TransportParameters

MaxAckDelay

MaxDatagramFrameSize

MaxIdleTimeout

MaxUdpPayloadSize

OriginalDestinationConnectionId

PreferredAddress

RetrySourceConnectionId

TransportParameters

ValidationError

ZeroRttParameters

Enums

MigrationSupport

Traits

TransportParameter

Trait for an transport parameter value

TransportParameterValidator

Trait for validating transport parameter values

Functions

compute_data_window

Computes a data window with the given configuration

Type Definitions

ClientTransportParameters

Specific TransportParameters sent by the client endpoint

ServerTransportParameters

Specific TransportParameters sent by the server endpoint