Struct s2n_quic_core::varint::VarInt

pub struct VarInt(_);

Implementations

impl VarInt

pub const MAX: Self = _

pub const ZERO: Self = _

pub fn new(v: u64) -> Result<Self, VarIntError>

pub const unsafe fn new_unchecked(value: u64) -> Self

Returns a VarInt without validating the value is less than VarInt::MAX

Safety

Callers need to ensure the value is less than or equal to VarInt::MAX

pub const fn from_u8(v: u8) -> Self

pub const fn from_u16(v: u16) -> Self

pub const fn from_u32(v: u32) -> Self

pub const fn as_u64(self) -> u64

pub fn checked_add(self, value: Self) -> Option<Self>

pub fn checked_add_usize(self, value: usize) -> Option<Self>

pub fn saturating_add(self, value: Self) -> Self

pub fn checked_sub(self, value: Self) -> Option<Self>

pub fn saturating_sub(self, value: Self) -> Self

pub fn checked_mul(self, value: Self) -> Option<Self>

pub fn saturating_mul(self, value: Self) -> Self

pub fn checked_div(self, value: Self) -> Option<Self>

pub fn encode_updated<E: Encoder>(self, replacement: Self, encoder: &mut E)

Re-encodes a replacement value where self was used as a placeholder.